srp: refactor groups and errors
Factors apart the junk drawer `types` module, moving the `SrpGroup` type
into the `groups` module, and renaming the remaining `types` module into
an `errors` module.
