+++ srp/src/groups.rs
@@ -4,9 +4,29 @@
 //! groups. Additionally, it is not recommended to use `G_1024` and `G_1536`,
 //! they are provided only for compatibility with the legacy software.
 
-use crate::types::SrpGroup;
-use crypto_bigint::BoxedUint;
+use crypto_bigint::{
+    BoxedUint, Odd, Resize,
+    modular::{BoxedMontyForm, BoxedMontyParams},
+};
 use once_cell::sync::Lazy;
+
+/// Group used for SRP computations
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub struct SrpGroup {
+    /// A large safe prime (N = 2q+1, where q is prime)
+    pub n: BoxedMontyParams,
+    /// A generator modulo N
+    pub g: BoxedMontyForm,
+}
+
+impl SrpGroup {
+    /// Initialize a new group from the given boxed integers.
+    pub fn new(n: BoxedUint, g: BoxedUint) -> Self {
+        let n = BoxedMontyParams::new(Odd::new(n).expect("n should be odd"));
+        let g = BoxedMontyForm::new(g.resize(n.bits_precision()), &n);
+        Self { n, g }
+    }
+}
 
 pub static G_1024: Lazy<SrpGroup> = Lazy::new(|| {
     SrpGroup::new(
@@ -43,16 +63,15 @@ pub static G_4096: Lazy<SrpGroup> = Lazy::new(|| {
     )
 });
 
-pub static G_6144: Lazy<SrpGroup> = Lazy::new(|| {
-    SrpGroup::new(
-        BoxedUint::from_be_slice_vartime(include_bytes!("groups/6144.bin")),
-        BoxedUint::from_be_slice_vartime(&[5]),
-    )
-});
+#[cfg(test)]
+mod tests {
+    use crate::groups::G_1024;
+    use crate::utils::compute_k;
+    use sha1::Sha1;
 
-pub static G_8192: Lazy<SrpGroup> = Lazy::new(|| {
-    SrpGroup::new(
-        BoxedUint::from_be_slice_vartime(include_bytes!("groups/8192.bin")),
-        BoxedUint::from_be_slice_vartime(&[19]),
-    )
-});
+    #[test]
+    fn test_k_1024_sha1() {
+        let k = compute_k::<Sha1>(&G_1024).to_be_bytes_trimmed_vartime();
+        assert_eq!(&*k, include_bytes!("test/k_sha1_1024.bin"));
+    }
+}
