+++ srp/src/lib.rs
@@ -58,7 +58,9 @@
 extern crate alloc;
 
 pub mod client;
+pub mod errors;
 pub mod groups;
 pub mod server;
-pub mod types;
 pub mod utils;
+
+pub use groups::SrpGroup;
